# src/main/resources/api.yaml
swagger: "2.0"
info:
  description: "This is a sample server power of attorney service"
  version: "1.0.0"
  title: "Power of attorney"
host: "localhost:8080"
tags:
- name: "Power of Attorney"
  description: "Everything about your power of attorney"
- name: "Debit Card"
  description: "Access to debit card details"
- name: "Credit Card"
  description: "Access to credit card details"
schemes:
- "http"
paths:
  /power-of-attorneys:
    get:
      tags:
      - "Power of Attorney"
      summary: "Get all power of attorneys"
      description: "Provides list of power of attorneys for current user"
      operationId: "getAllPowerOfAttorneys"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/PowerOfAttorneyReference"
  /power-of-attorneys/{id}:
    get:
      tags:
      - "Power of Attorney"
      summary: "Get Detail of Power of Attorney"
      description: "Get Detail of given Power of Attorney"
      operationId: "getPowerOfAttorneyDetail"
      parameters:
        - in: path
          name: id   # Note the name is the same as in the path
          required: true
          type: "string"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/PowerOfAttorney"
        404:
          description: "Power of Attorney not found"
  /debit-cards/{id}:
    get:
      tags:
      - "Debit Card"
      summary: "Get Detail of debit card"
      description: "Get Detail of given debit card"
      operationId: "getDebitCardDetail"
      parameters:
        - in: path
          name: id   # Note the name is the same as in the path
          required: true
          type: "string"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/DebitCard"
        404:
          description: "Debit card not found"
  /credit-cards/{id}:
    get:
      tags:
      - "Credit Card"
      summary: "Get Detail of credit card"
      description: "Get Detail of given credit card"
      operationId: "getDebitCardDetail"
      parameters:
        - in: path
          name: id   # Note the name is the same as in the path
          required: true
          type: "string"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/CreditCard"
        404:
          description: "Credit card not found"
definitions:
  PowerOfAttorneyReference:
    type: "object"
    properties:
      id:
        type: "string"
  PowerOfAttorney:
    type: "object"
    properties:
      id:
        type: "string"
      grantor:
        type: "string"
      grantee:
        type: "string"
      account:
        type: "string"
      direction:
        $ref: "#/definitions/Direction"
      authorizations:
        type: "array"
        items:
          $ref: "#/definitions/Authorization"
      cards:
        type: "array"
        items:
          $ref: "#/definitions/CardReference"
  CardReference:
    type: "object"
    properties:
      id:
        type: "string"
      type:
         $ref: "#/definitions/CardType"
  DebitCard:
    type: "object"
    properties:
      id:
        type: "string"
      cardNumber:
        type: "integer"
      sequenceNumber:
        type: "integer"
      cardHolder:
        type: "string"
      atmLimit:
        $ref: "#/definitions/Limit"
      posLimit:
        $ref: "#/definitions/Limit"
      contactless:
        type: "boolean"
  CreditCard:
    type: "object"
    properties:
      id:
        type: "string"
      cardNumber:
        type: "integer"
      sequenceNumber:
        type: "integer"
      cardHolder:
        type: "string"
      monhtlyLimit:
        type: "integer"
  Limit:
    type: "object"
    properties:
      limit:
        type: "integer"
      periodUnit:
         $ref: "#/definitions/PeriodUnit"
  Authorization:
    type: "string"
    enum: ["DEBIT_CARD","CREDIT_CARD","VIEW","PAYMENT"]
  Direction:
    type: "string"
    enum: ["GIVEN", "RECEIVED"]
  CardType:
    type: "string"
    enum: ["DEBIT_CARD", "CREDIT_CARD"]
  PeriodUnit:
    type: "string"
    enum: ["PER_DAY","PER_WEEK","PER_MONTH"]